requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
CurrencyConverter==0.17.23
Django==4.2.5
django-admin-interface==0.28.6
django-colorfield==0.11.0
django-cors-headers==4.2.0
django-crispy-forms==2.1
django-filter==23.3
django-paypal==2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
fastboot==0.0.3
filelock==3.12.4
flake8==7.0.0
Flask==2.1.2
forex-python==1.8
frozendict==2.4.4
future==0.18.3
gevent==23.9.1
greenlet==3.0.0
html5lib==1.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
json5==0.9.24
lxml==5.2.2
Markdown==3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mock==5.1.0
multitasking==0.0.11
numpy==2.0.0
pandas==2.2.2
peewee==3.17.5
Pillow==10.0.1
platformdirs==4.2.2
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pyflakes==3.2.0
pypng==0.20220715.0
pyserial==3.5
PySide6==6.6.3.1
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2023.3.post1
pyusb==1.2.1
pywin32==306
qrcode==7.4.2
requests==2.31.0
requests-file==1.5.1
shiboken6==6.6.3.1
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
tldextract==3.6.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
webencodings==0.5.1
Werkzeug==3.0.3
yfinance==0.2.40
zope.event==5.0
zope.interface==6.0
